third_party/WebKit/LayoutTests/platform/mac/external/wpt/css/css-tables/table-model-fixup-2-expected.txt

This is a testharness.js-based test.
PASS Replaced elements inside a table cannot be table-row and are considered block -- input elements
PASS Replaced elements inside a table cannot be table-row and are considered block -- img elements
PASS Replaced elements inside a table cannot be table-column and are considered inline -- input elements
PASS Replaced elements inside a table cannot be table-column and are considered inline -- img elements
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- input elements
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- img elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=button elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=file elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- img elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=text elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=button elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=file elements
FAIL Replaced elements outside a table cannot be table and are considered block -- img elements assert_not_equals: got disallowed value 2046
FAIL Replaced elements outside a table cannot be table-row and are considered block -- input=text elements assert_not_equals: got disallowed value 2142
FAIL Replaced elements outside a table cannot be table-row and are considered block -- input=button elements assert_not_equals: got disallowed value 2257
PASS Replaced elements outside a table cannot be table-row and are considered block -- input=file elements
FAIL Replaced elements outside a table cannot be table-row and are considered block -- img elements assert_not_equals: got disallowed value 2504
FAIL Replaced elements outside a table cannot be table-row-group and are considered block -- input=text elements assert_not_equals: got disallowed value 2600
FAIL Replaced elements outside a table cannot be table-row-group and are considered block -- input=button elements assert_not_equals: got disallowed value 2715
PASS Replaced elements outside a table cannot be table-row-group and are considered block -- input=file elements
FAIL Replaced elements outside a table cannot be table-row-group and are considered block -- img elements assert_not_equals: got disallowed value 2962
PASS Replaced elements outside a table cannot be table-column and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-column and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-column and are considered inline -- input=file elements assert_equals: expected 3306 but got 3288
PASS Replaced elements outside a table cannot be table-column and are considered inline -- img elements
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-cell and are considered inline -- input=file elements assert_equals: expected 3764 but got 3746
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- img elements
Harness: the test ran to completion.
